Updates User Controller, Model and Repository
in userController.js:
- Removes unnecessary try/catch in index due no error could happend.
- Delete method now throws an error if the received id is undefined.

in userModel.js:
model now returns UserModel, being able to test its functionality with jest creating a sequelize instance in memory and synchronizing the model without using the real database

in userRepository.js:
- Now, getById method throws an error if the received id is undefined, also if sequelize couldn't find anything with findByPk, it returns null, not undefined, so fixes the if statement to cover that path
- Finally, delete method now throws an error if the received id is undefined, removing unnecessary deleteUser variable and statement.

+++ src/module/user/controller/userController.js
@@ -15,17 +15,12 @@ module.exports = class UserController extends AbstractController{
 		app.get(`${ROUTE}/create`, this.create.bind(this));
 		app.post(`${ROUTE}/create`, this.save.bind(this));
 		app.post(`${ROUTE}/delete/:id`, this.delete.bind(this)); 
-		app.get(`${ROUTE}/edit/:id`, this.edit.bind(this));     
+		app.get(`${ROUTE}/edit/:id`, this.edit.bind(this)); 
 	}
 
 	async index(req, res){
-		try{
-			const users = await this.userService.getAll();
-			res.render("user/view/index.html", { users: users ? users : false, pageTitle: "Users list" });
-		}catch(e){
-			req.session.errors = [e.message, e.stack];
-			res.redirect("/");
-		}
+		const users = await this.userService.getAll();
+		res.render("user/view/index.html", { users: users ? users : false, pageTitle: "Users list" });
 	}
 
 	create(req, res){
@@ -59,8 +54,13 @@ module.exports = class UserController extends AbstractController{
 	}
 
 	async delete(req, res){
+		const userId = req.params.id;
+
+		if(userId === undefined){
+			throw new UserIdNotDefinedError("User id is not defined");
+		}
+
 		try{
-			const userId = req.params.id;
 			const deletedUser = await this.userService.delete(userId);
 
 			req.session.messages = [`User #${deletedUser.id} ${deletedUser.names} ${deletedUser.surnames} deleted succesfully`];
